🛠️ Refactor: Enhance structure and enforce error handling conventions
- Added centralized error reporting via `src/reporter/reporter.go` (Singleton pattern, SRP). Avoided `utils` folder to reduce sparsity.
- Refactored `requestData` to use `http.NewRequest`, eliminating custom `rcwrap` (Dead Code removal per Fowler).
- Removed ghost code from `api/selichoje.go` and `api/selichoje_test.go` (Reducing cognitive load).
- Hidden raw error exposure using `http.Error` instead of custom header/body writing for 500s.
- Updated deprecated `ioutil.ReadAll` to `io.ReadAll`.
- Ran `go mod tidy` to clean up unused dependencies, fixing CI build issues.
- Updated `go.mod` to `go 1.22.1` to match `mise.toml`.
- Added `mise` and `mise.local.toml` to `.gitignore`.
