+++ .gitignore
@@ -1,1 +1,3 @@
 .vercel
+mise
+mise.local.toml
