+++ go.mod
@@ -1,3 +1,3 @@
 module github.com/lucasew/bcb-selic-hoje
 
-go 1.16
+go 1.22.1
